Update menu.py
In the _ctl_submenu function of adding parentCTL/childCTL, the rigbits module of mgear is not imported, which leads to an error message. The corresponding control cannot be created. It can only be created correctly after clicking other menus. For this reason, I added "from mgear import rigbits". Code and make it execute correctly.
The error description is in the forum http://forum.mgear-framework.com/t/mgear-3-7-8-rigbits-added-parent-child-ctl-error/2051